Added functionality for LIC2
Added LIC 2 to close #4

LIC2 description:
There exists at least one set of three consecutive data points which form an angle such that:
angle < (PI − EPSILON)
or
angle > (PI + EPSILON)
The second of the three consecutive points is always the vertex of the angle. If either the first
point or the last point (or both) coincides with the vertex, the angle is undefined and the LIC
is not satisfied by those three points.
(0 ≤ EPSILON < PI