Add the Image File used in README
The image is an jpg which is used as the cover of our README. It is generated by Adobe Firefly with the prompt "The Launch Interceptor Project implements a function called DECIDE() that determines whether an interceptor should be launched based on input radar tracking information. for anti-ballistic missile systems, providing a systematic approach to decision-making based on real-time radar data. 4K, sci-fi"